+++ planner_cspace/test/src/test_cyclic_vec.cpp
@@ -27,11 +27,12 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
+#include <cmath>
 #include <cstddef>
 
-#include <gtest/gtest.h>
-
 #include <planner_cspace/cyclic_vec.h>
+
+#include <gtest/gtest.h>
 
 TEST(CyclicVec, InitFloat)
 {
@@ -104,7 +105,7 @@ TEST(CyclicVec, LengthInt)
 
   ASSERT_EQ(v.sqlen(), 25);
   ASSERT_EQ(v.len(), 5.0);
-  ASSERT_EQ(v.norm(), sqrtf(50));
+  ASSERT_EQ(v.norm(), std::sqrt(50.0f));
 }
 
 TEST(CyclicVec, LengthFloat)
@@ -114,7 +115,7 @@ TEST(CyclicVec, LengthFloat)
 
   ASSERT_EQ(v.sqlen(), 25.0);
   ASSERT_EQ(v.len(), 5.0);
-  ASSERT_EQ(v.norm(), sqrtf(50.0));
+  ASSERT_EQ(v.norm(), std::sqrt(50.0f));
 }
 
 TEST(CyclicVec, Cycle)
